requirements.txt
crispy-bootstrap4==2022.1
Django==4.2.1
django-auto-logout==0.5.1
django-bootstrap-datepicker-plus==4.0.0
django-cleanup==7.0.0
django-crispy-forms==2.0
django-debug-toolbar==3.2.2
django-extensions==3.2.3
django-filter==23.2
django-htmx==1.14.0
django-import-export==3.2.0
django-pwa==1.1.0
django-select2==8.1.2
django-storages==1.13.2
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.5.0
pymemcache==4.0.0
python-dateutil==2.8.2
python-memcached==1.59
requests==2.31.0
shortuuid==1.0.11

